pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doctor-scribe"
version = "0.1.0"
description = "Clinical note fine-tuning with QLoRA"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "evaluate>=0.4.0",
    "rouge-score>=0.1.2",
    "bert-score>=0.3.13",
    "openai>=1.12.0",
    "google-generativeai>=0.4.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "mlflow>=2.11.0",
    "hydra-core>=1.3.0",
    "tiktoken>=0.6.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
doctor-scribe-serve = "scripts.serve:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--cov=src", "--cov-report=term-missing"]
asyncio_mode = "auto"
